fix shortcut registration
